# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Battleship API",
    "description": "PvP Battleship game REST API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "tags": ["Health"],
        "responses": {
          "200": {
            "description": "Server is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/games": {
      "post": {
        "summary": "Create a new game",
        "tags": ["Games"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Player name",
                    "example": "Player1"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Game created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/games/{code}/join": {
      "post": {
        "summary": "Join an existing game",
        "tags": ["Games"],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "6-character game code",
            "example": "ABC123"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Player name",
                    "example": "Player2"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Joined game successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameCreatedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Game not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/games/{code}": {
      "get": {
        "summary": "Get game state",
        "tags": ["Games"],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "6-character game code"
          },
          {
            "name": "X-Player-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Player token received from create/join"
          }
        ],
        "responses": {
          "200": {
            "description": "Game state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameState"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing player token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/games/{code}/place-ships": {
      "post": {
        "summary": "Place ships on the board",
        "tags": ["Gameplay"],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "6-character game code"
          },
          {
            "name": "X-Player-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Player token"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ships"],
                "properties": {
                  "ships": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Ship"
                    },
                    "minItems": 5,
                    "maxItems": 5
                  }
                }
              },
              "example": {
                "ships": [
                  {
                    "name": "Carrier",
                    "size": 5,
                    "positions": [
                      {"x": 0, "y": 0},
                      {"x": 1, "y": 0},
                      {"x": 2, "y": 0},
                      {"x": 3, "y": 0},
                      {"x": 4, "y": 0}
                    ]
                  },
                  {
                    "name": "Battleship",
                    "size": 4,
                    "positions": [
                      {"x": 0, "y": 1},
                      {"x": 1, "y": 1},
                      {"x": 2, "y": 1},
                      {"x": 3, "y": 1}
                    ]
                  },
                  {
                    "name": "Cruiser",
                    "size": 3,
                    "positions": [
                      {"x": 0, "y": 2},
                      {"x": 1, "y": 2},
                      {"x": 2, "y": 2}
                    ]
                  },
                  {
                    "name": "Submarine",
                    "size": 3,
                    "positions": [
                      {"x": 0, "y": 3},
                      {"x": 1, "y": 3},
                      {"x": 2, "y": 3}
                    ]
                  },
                  {
                    "name": "Destroyer",
                    "size": 2,
                    "positions": [
                      {"x": 0, "y": 4},
                      {"x": 1, "y": 4}
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ships placed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Ships placed successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ship placement",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/games/{code}/fire": {
      "post": {
        "summary": "Fire at a coordinate",
        "tags": ["Gameplay"],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "6-character game code"
          },
          {
            "name": "X-Player-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Player token"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["x", "y"],
                "properties": {
                  "x": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9,
                    "description": "X coordinate (0-9)"
                  },
                  "y": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9,
                    "description": "Y coordinate (0-9)"
                  }
                }
              },
              "example": {
                "x": 5,
                "y": 3
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Fire result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FireResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid move or not your turn",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Coordinate": {
        "type": "object",
        "required": ["x", "y"],
        "properties": {
          "x": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9
          },
          "y": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9
          }
        }
      },
      "Ship": {
        "type": "object",
        "required": ["name", "size", "positions"],
        "properties": {
          "name": {
            "type": "string",
            "enum": ["Carrier", "Battleship", "Cruiser", "Submarine", "Destroyer"]
          },
          "size": {
            "type": "integer",
            "enum": [5, 4, 3, 2]
          },
          "positions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Coordinate"
            }
          }
        }
      },
      "Shot": {
        "type": "object",
        "properties": {
          "x": {
            "type": "integer"
          },
          "y": {
            "type": "integer"
          },
          "hit": {
            "type": "boolean"
          }
        }
      },
      "GameCreatedResponse": {
        "type": "object",
        "properties": {
          "gameCode": {
            "type": "string",
            "example": "ABC123"
          },
          "gameId": {
            "type": "string",
            "format": "uuid"
          },
          "playerToken": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": ["waiting", "placing", "playing", "finished"]
          }
        }
      },
      "GameState": {
        "type": "object",
        "properties": {
          "game": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "code": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": ["waiting", "placing", "playing", "finished"]
              },
              "current_turn": {
                "type": "string",
                "enum": ["player1", "player2"],
                "nullable": true
              },
              "winner": {
                "type": "string",
                "enum": ["player1", "player2"],
                "nullable": true
              }
            }
          },
          "you": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "board": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Ship"
                }
              },
              "shots": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Shot"
                }
              },
              "ready": {
                "type": "boolean"
              }
            }
          },
          "opponent": {
            "type": "object",
            "nullable": true,
            "properties": {
              "name": {
                "type": "string"
              },
              "shots": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Shot"
                }
              },
              "ready": {
                "type": "boolean"
              },
              "shipsSunk": {
                "type": "integer"
              }
            }
          },
          "isYourTurn": {
            "type": "boolean"
          }
        }
      },
      "FireResult": {
        "type": "object",
        "properties": {
          "hit": {
            "type": "boolean",
            "description": "Whether the shot hit a ship"
          },
          "sunk": {
            "type": "string",
            "nullable": true,
            "description": "Name of ship sunk, if any"
          },
          "gameOver": {
            "type": "boolean",
            "description": "Whether the game has ended"
          },
          "winner": {
            "type": "string",
            "enum": ["player1", "player2"],
            "nullable": true,
            "description": "Winner if game is over"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Health",
      "description": "Health check endpoints"
    },
    {
      "name": "Games",
      "description": "Game management endpoints"
    },
    {
      "name": "Gameplay",
      "description": "Gameplay endpoints"
    }
  ]
}